research/requirements.txt
# Core data processing
pandas>=2.0.0
numpy>=1.24.0

# API clients
requests>=2.31.0
python-dotenv>=1.0.0

# AI/LLM clients (choose one or both)
openai>=1.0.0
anthropic>=0.21.0

# Data analysis and visualization
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# Database (for later phases)
psycopg2-binary>=2.9.0
sqlalchemy>=2.0.0

# Utilities
tqdm>=4.65.0  # Progress bars
tenacity>=8.2.0  # Retry logic for API calls
